+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - New `LegacyPolygons` parameter for Geo Indexes
 - New parameters (`cacheEnabled` and `storedValues`) for Persistent Indexes
 - New analyzers: `classification`, `nearest neighbors`, `minhash`
+- Add support for Inverted index
 
 ## [1.3.3](https://github.com/arangodb/go-driver/tree/v1.3.3) (2022-07-27)
 - Fix `lastValue` field type